# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platinum Axe API",
    "description": "\u682a\u5f0f\u6295\u8cc7AI\u30af\u30a9\u30f3\u30c4\u5206\u6790\u30b7\u30b9\u30c6\u30e0 - Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/history/latest": {
      "get": {
        "tags": [
          "history",
          "history"
        ],
        "summary": "Get Latest Results",
        "description": "\u76f4\u8fd1\u306e\u30e9\u30a6\u30f3\u30c9\u7d50\u679c\u3092\u53d6\u5f97\uff08BUY/SELL\uff09\n\nArgs:\n    index_filter: all\uff08\u5168\u9298\u67c4\uff09 | nikkei225\uff08\u65e5\u7d4c225\u306e\u307f\uff09 | topix\uff08TOPIX\u306e\u307f\uff09\n    db: DB\u30bb\u30c3\u30b7\u30e7\u30f3\n\nReturns:\n    \u76f4\u8fd1\u306eBUY/SELL\u30e9\u30a6\u30f3\u30c9\u7d50\u679c",
        "operationId": "get_latest_results_api_v1_history_latest_get",
        "parameters": [
          {
            "name": "index_filter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "nikkei225",
                "topix"
              ],
              "type": "string",
              "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc",
              "default": "all",
              "title": "Index Filter"
            },
            "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/summary": {
      "get": {
        "tags": [
          "history",
          "history"
        ],
        "summary": "Get Performance Summary",
        "description": "\u5168\u4f53\u306e\u30d1\u30d5\u30a9\u30fc\u30de\u30f3\u30b9\u30b5\u30de\u30ea\u30fc\u3092\u53d6\u5f97\n\nArgs:\n    index_filter: all\uff08\u5168\u9298\u67c4\uff09 | nikkei225\uff08\u65e5\u7d4c225\u306e\u307f\uff09 | topix\uff08TOPIX\u306e\u307f\uff09\n    db: DB\u30bb\u30c3\u30b7\u30e7\u30f3\n\nReturns:\n    BUY/SELL\u306e\u5168\u4f53\u30d1\u30d5\u30a9\u30fc\u30de\u30f3\u30b9\u7d71\u8a08",
        "operationId": "get_performance_summary_api_v1_history_summary_get",
        "parameters": [
          {
            "name": "index_filter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "nikkei225",
                "topix"
              ],
              "type": "string",
              "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc",
              "default": "all",
              "title": "Index Filter"
            },
            "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "history",
          "history"
        ],
        "summary": "Get Round History",
        "description": "\u30e9\u30a6\u30f3\u30c9\u5c65\u6b74\u3092\u53d6\u5f97\uff08\u30da\u30fc\u30b8\u30cd\u30fc\u30b7\u30e7\u30f3\u4ed8\u304d\uff09\n\nArgs:\n    round_type: BUY | SELL | None\uff08\u5168\u3066\uff09\n    index_filter: all\uff08\u5168\u9298\u67c4\uff09 | nikkei225\uff08\u65e5\u7d4c225\u306e\u307f\uff09 | topix\uff08TOPIX\u306e\u307f\uff09\n    page: \u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09\n    limit: 1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570\uff081-100\uff09\n    db: DB\u30bb\u30c3\u30b7\u30e7\u30f3\n\nReturns:\n    \u30e9\u30a6\u30f3\u30c9\u5c65\u6b74\u3068\u30da\u30fc\u30b8\u30cd\u30fc\u30b7\u30e7\u30f3\u60c5\u5831",
        "operationId": "get_round_history_api_v1_history_get",
        "parameters": [
          {
            "name": "round_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "BUY",
                    "SELL"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "\u30e9\u30a6\u30f3\u30c9\u30bf\u30a4\u30d7\uff08\u672a\u6307\u5b9a\u3067\u5168\u3066\uff09",
              "title": "Round Type"
            },
            "description": "\u30e9\u30a6\u30f3\u30c9\u30bf\u30a4\u30d7\uff08\u672a\u6307\u5b9a\u3067\u5168\u3066\uff09"
          },
          {
            "name": "index_filter",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "nikkei225",
                "topix"
              ],
              "type": "string",
              "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc",
              "default": "all",
              "title": "Index Filter"
            },
            "description": "\u6307\u6570\u30d5\u30a3\u30eb\u30bf\u30fc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "\u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09",
              "default": 1,
              "title": "Page"
            },
            "description": "\u30da\u30fc\u30b8\u756a\u53f7\uff081\u59cb\u307e\u308a\uff09"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570",
              "default": 10,
              "title": "Limit"
            },
            "description": "1\u30da\u30fc\u30b8\u3042\u305f\u308a\u306e\u4ef6\u6570"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history/{round_id}": {
      "get": {
        "tags": [
          "history",
          "history"
        ],
        "summary": "Get Round Detail",
        "description": "\u7279\u5b9a\u30e9\u30a6\u30f3\u30c9\u306e\u8a73\u7d30\u3092\u53d6\u5f97\uff08\u63a8\u5968\u9298\u67c4 + \u7d50\u679c\u30c7\u30fc\u30bf\u4ed8\u304d\uff09\n\nArgs:\n    round_id: \u30e9\u30a6\u30f3\u30c9ID\uff08\u4f8b: 2026-W29-BUY\uff09\n    db: DB\u30bb\u30c3\u30b7\u30e7\u30f3\n\nReturns:\n    \u30e9\u30a6\u30f3\u30c9\u8a73\u7d30 + \u63a8\u5968\u9298\u67c4 + \u7d50\u679c\u30c7\u30fc\u30bf\n\nRaises:\n    HTTPException: \u30e9\u30a6\u30f3\u30c9\u304c\u898b\u3064\u304b\u3089\u306a\u3044\u5834\u5408\uff08404\uff09",
        "operationId": "get_round_detail_api_v1_history__round_id__get",
        "parameters": [
          {
            "name": "round_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Round Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}